test: Consolidate all security group tests
